+++ src/Brassica/SoundChange/Apply.hs
@@ -22,6 +22,7 @@ module Brassica.SoundChange.Apply
        , applyChanges
        , Log(..)
        , LogItem(..)
+       , rulesNotApplied
        -- * Reporting results
        , getOutput
        , getReports